add i18n entries and eslint fixes
