+++ pkg/sbombastic-image-vulnerability-scanner/constants/index.ts
@@ -1,2 +1,1 @@
-export * from './scan-interval-options';
 export * from './securityConstants';
