+++ pkg/sbombastic-image-vulnerability-scanner/l10n/en-us.yaml
@@ -87,9 +87,13 @@ imageScanner:
         scan:
           label: Scanning
           repoName: Repositories to scan
-          placeholder: Leave empty to scan all. Type and press Enter to add repositories.
+          placeholder:
+            dockerhub: Leave empty to scan all. Type and press Enter to add repositories.
+            nocatalog: Add at least one repository. Type and press Enter to add.
           schedule:
             label: Scan Interval
+            manualScan: Manual scan
+            placeholder: Selecting '{manualScan}' disables automatic scanning
       scan: Start Scan
       meta:
         namespace: Namespace
